Moved DataStreamsCheckpointer to experimental API
