+++ dd-trace-api/build.gradle
@@ -23,8 +23,9 @@ excludedClassesCoverage += [
   'datadog.trace.api.experimental.ProfilingScope',
   'datadog.trace.api.experimental.ProfilingContext',
   'datadog.trace.api.experimental.ProfilingContextSetter.NoOp',
-  'datadog.trace.api.datastreams.DataStreamsCheckpointer',
-  'datadog.trace.api.datastreams.DataStreamsContextCarrier',
+  'datadog.trace.api.experimental.DataStreamsCheckpointer',
+  'datadog.trace.api.experimental.DataStreamsCheckpointer.NoOp',
+  'datadog.trace.api.experimental.DataStreamsContextCarrier',
   'datadog.appsec.api.blocking.*',
 ]
