+++ dd-trace-api/src/main/java/datadog/trace/api/experimental/DataStreamsCheckpointer.java
@@ -1,7 +1,20 @@
-package datadog.trace.api.datastreams;
+package datadog.trace.api.experimental;
+
+import datadog.trace.api.GlobalTracer;
+import datadog.trace.api.Tracer;
+import datadog.trace.api.internal.InternalTracer;
 
 /** An interface to Data Streams checkpointer, allowing passing the context manually. */
 public interface DataStreamsCheckpointer {
+  static DataStreamsCheckpointer get() {
+    Tracer tracer = GlobalTracer.get();
+    if (tracer instanceof InternalTracer) {
+      return ((InternalTracer) tracer).getDataStreamsCheckpointer();
+    }
+
+    return NoOp.INSTANCE;
+  }
+
   /**
    * @param type The type of the checkpoint, usually the streaming technology being used. Examples:
    *     kafka, kinesis, sns etc.
@@ -20,4 +33,19 @@ public interface DataStreamsCheckpointer {
    *     wrapper around message headers.
    */
   void setProduceCheckpoint(String type, String target, DataStreamsContextCarrier carrier);
+}
+
+final class NoOp implements DataStreamsCheckpointer {
+
+  public static final DataStreamsCheckpointer INSTANCE = new NoOp();
+
+  @Override
+  public void setConsumeCheckpoint(String type, String source, DataStreamsContextCarrier carrier) {
+
+  }
+
+  @Override
+  public void setProduceCheckpoint(String type, String target, DataStreamsContextCarrier carrier) {
+
+  }
 }
