+++ dd-trace-core/src/main/java/datadog/trace/core/CoreTracer.java
@@ -31,8 +31,8 @@ import datadog.trace.api.IdGenerationStrategy;
 import datadog.trace.api.StatsDClient;
 import datadog.trace.api.TracePropagationStyle;
 import datadog.trace.api.config.GeneralConfig;
-import datadog.trace.api.datastreams.DataStreamsCheckpointer;
-import datadog.trace.api.datastreams.DataStreamsContextCarrier;
+import datadog.trace.api.experimental.DataStreamsCheckpointer;
+import datadog.trace.api.experimental.DataStreamsContextCarrier;
 import datadog.trace.api.experimental.Profiling;
 import datadog.trace.api.gateway.CallbackProvider;
 import datadog.trace.api.gateway.InstrumentationGateway;
