+++ dd-trace-ot/build.gradle
@@ -109,7 +109,6 @@ shadowJar {
     exclude('datadog.opentracing.resolver.*')
     exclude('datadog.trace.api.*')
     exclude('datadog.trace.api.config.*')
-    exclude('datadog.trace.api.datastreams.*')
     exclude('datadog.trace.api.experimental.*')
     exclude('datadog.trace.api.interceptor.*')
     exclude('datadog.trace.api.internal.*')
