+++ dd-trace-ot/src/main/java/datadog/opentracing/DDTracer.java
@@ -4,7 +4,7 @@ import datadog.trace.api.Config;
 import datadog.trace.api.DDTags;
 import datadog.trace.api.GlobalTracer;
 import datadog.trace.api.StatsDClient;
-import datadog.trace.api.datastreams.DataStreamsCheckpointer;
+import datadog.trace.api.experimental.DataStreamsCheckpointer;
 import datadog.trace.api.experimental.Profiling;
 import datadog.trace.api.interceptor.TraceInterceptor;
 import datadog.trace.api.internal.InternalTracer;
